src/java/br/com/treebu/model/Autor.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package br.com.treebu.model;

import java.util.Date;

/**
 *
 * @author vinicius caetano
 */
public class Autor {
    
    private int codigo;
    private String nome;
    private String nacionalidade;
    private Date dataNascimento;
    private boolean status;
    private String descricao;

    public Autor() {
    }

    public Autor(int codigo, String nome, String nacionalidade, Date dataNascimento, boolean status, String descricao) {
        this.codigo = codigo;
        this.nome = nome;
        this.nacionalidade = nacionalidade;
        this.dataNascimento = dataNascimento;
        this.status = status;
        this.descricao = descricao;
    }

    public int getCodigo() {
        return codigo;
    }

    public void setCodigo(int codigo) {
        this.codigo = codigo;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public String getNacionalidade() {
        return nacionalidade;
    }

    public void setNacionalidade(String nacionalidade) {
        this.nacionalidade = nacionalidade;
    }

    public Date getDataNascimento() {
        return dataNascimento;
    }

    public void setDataNascimento(Date dataNascimento) {
        this.dataNascimento = dataNascimento;
    }

    public boolean isStatus() {
        return status;
    }

    public void setStatus(boolean status) {
        this.status = status;
    }

    public String getDescricao() {
        return descricao;
    }

    public void setDescricao(String descricao) {
        this.descricao = descricao;
    }

    @Override
    public String toString() {
        return "Autor{" + "codigo=" + codigo + ", nome=" + nome + ", nacionalidade=" + nacionalidade + ", dataNascimento=" + dataNascimento + ", status=" + status + ", descricao=" + descricao + '}';
    }
    
  
}
